use GemQuoteAsset
